Modify approach to serving UI
Instead of combining all of the UI objects into a single page that is
served to all of the windows, make the main UI a dynamic function that
knows what window it's in, and serves only the appropriate UI.

Also, dependencies no longer need to be provided explicitly. This was
only necessary before due to `paste`-ing of UI objects, which should
generally never be done because it causes dependency information to
be dropped on the floor. Instead, we compose our UI using tag objects
all the way up to the page level.