mwshiny v2.1.1 (server redundancy bugfix)
serv_calc() functions are duplicated for each window in <=v2.1.0. This sort of fixes this, but in order for it to work, you have to keep the first rendered WindowSelector window open for the duration of app use. Definitely is not a permanent solution, and will not be put on CRAN. I would recommend using the CRAN version, which is stable, if you don't have any big computations.